Remove env variables

+++ cloudflare/removeDNSRecord.js
@@ -9,16 +9,22 @@ async function removeDNSRecord() {
     const ZONE_ID = process.env.ZONE_ID;
     const DNS_ID = process.env.DNS_ID;
     const UMAMI_DNS_ID = process.env.UMAMI_DNS_ID;
-
-    const url = `https://api.cloudflare.com/client/v4/zones/${ZONE_ID}/dns_records/${DNS_ID}`;
     const headers = {
       "X-Auth-Email": EMAIL,
       "X-Auth-Key": API_KEY,
       "Content-Type": "application/json",
     };
 
-    await axios.delete(url, { headers });
-    console.log("Certificate validation record deleted")
+    if (!DNS_ID && !UMAMI_DNS_ID) {
+      console.log('There are no DNS records to remove');
+    }
+
+    if (DNS_ID) {
+      const url = `https://api.cloudflare.com/client/v4/zones/${ZONE_ID}/dns_records/${DNS_ID}`;
+      await axios.delete(url, { headers });
+      console.log("Certificate validation record deleted")
+    }
+
     if (UMAMI_DNS_ID) {
       const url2 = `https://api.cloudflare.com/client/v4/zones/${ZONE_ID}/dns_records/${UMAMI_DNS_ID}`;
       await axios.delete(url2, { headers });
